Rationalised HatPtr and Types.  Moved from experiments to hat
