+++ hat/hat/src/main/java/hat/optools/FuncOpWrapper.java
@@ -24,9 +24,7 @@
  */
 package hat.optools;
 
-import hat.HatOps;
-import hat.KernelContext;
-import hat.buffer.Buffer;
+import hat.HatPtr;
 import hat.buffer.MappableIface;
 
 import java.lang.foreign.GroupLayout;
@@ -225,7 +223,7 @@ public class FuncOpWrapper extends OpWrapper<CoreOp.FuncOp> {
 
     public BiMap<Block.Parameter, CoreOp.VarOp> parameterVarOpMap = new BiMap<>();
     public BiMap<Block.Parameter, CoreOp.InvokeOp> parameterInvokeOpMap = new BiMap<>();
-    public BiMap<Block.Parameter, HatOps.HatPtrOp> parameterHatPtrOpMap = new BiMap<>();
+    public BiMap<Block.Parameter, HatPtr.HatPtrOp> parameterHatPtrOpMap = new BiMap<>();
     public FuncOpWrapper(CoreOp.FuncOp op) {
         super(op);
         op().parameters().forEach(parameter -> {
@@ -237,7 +235,7 @@ public class FuncOpWrapper extends OpWrapper<CoreOp.FuncOp> {
                     paramTable.add(Map.entry(parameter, varOp));
                 }else if (resultOp instanceof CoreOp.InvokeOp invokeOp) {
                     parameterInvokeOpMap.add(parameter,invokeOp);
-                }else if (resultOp instanceof HatOps.HatPtrOp hatPtrOp) {
+                }else if (resultOp instanceof HatPtr.HatPtrOp hatPtrOp) {
                     parameterHatPtrOpMap.add(parameter,hatPtrOp);
                 }else{
                     //System.out.println("neither varOp or an invokeOp "+resultOp.getClass().getName());
